Remove implicit conversion

+++ include/pybind11/pytypes.h
@@ -1078,7 +1078,7 @@ public:
     // Allow implicit conversion:
     bytearray(const std::string &s) : bytearray(s.data(), s.size()) { }
 
-    size_t size() const { return PyByteArray_Size(m_ptr); }
+    size_t size() const { return static_cast<size_t>(PyByteArray_Size(m_ptr)); }
 
     operator std::string() const {
         char *buffer = PyByteArray_AS_STRING(m_ptr);
